Moved LetterRack.java and LetterRackTest.java to proper folder
