+++ src/test/java/core_entities/game_parts/LetterRackTest.java
@@ -1,5 +1,7 @@
-package CoreEntities.GameParts;
+package core_entities.game_parts;
 
+import CoreEntities.GameParts.Bag;
+import CoreEntities.GameParts.Tile;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
